requirements.txt
python-dotenv
pandas 
numpy 
pymango
certifi
pymongo[srv]==4.16
scikit-learn

pyaml

# -e .
